requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
frozenlist==1.3.3
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.9
multidict==6.0.4
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.15
typing_extensions==4.5.0
yarl==1.9.2
